requirements.txt
# Core dependencies
python-dotenv>=1.0.0
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
altair>=5.0.0

# Database
sqlalchemy>=2.0.0
sqlmodel>=0.0.8

# Web scraping and API
requests>=2.28.0
beautifulsoup4>=4.12.0
selenium>=4.9.0
webdriver-manager>=3.8.0
finnhub-python>=2.4.0
alpha_vantage>=2.3.1

# NLP and processing
langchain>=0.0.267
anthropic>=0.5.0
nltk>=3.8.0
scikit-learn>=1.3.0
textblob>=0.17.0

# Date and time handling
pytz>=2023.3
python-dateutil>=2.8.2
